backend/requirements.txt
anyio==3.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
distro==1.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
fastapi==0.95.0
fatsecret==0.4.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
murmurhash==1.0.9
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pathy==0.10.1
preshed==3.0.8
pydantic==1.10.7
Pygments==2.14.0
python-dotenv==1.0.0
rauth==0.7.3
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.1
shell-gpt==0.7.3
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
starlette==0.26.1
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
wasabi==1.1.1
